requirements.txt
PySide6~=6.7
qrcode[pil,svg]~=7.4
Pillow~=10.4
opencv-python~=4.10
pyzbar~=0.1.8
